// ui/src/gen/ExperienceProposalView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ExperienceEntry } from "./ExperienceEntry";
import type { ExperienceRequest } from "./ExperienceRequest";
import type { ExperienceSource } from "./ExperienceSource";
import type { LegacyRange } from "./LegacyRange";

export type ExperienceProposalView = { proposal_id: string, request: ExperienceRequest, source: ExperienceSource, previous_entries: Array<ExperienceEntry>, legacy: LegacyRange | null, recovery_pending: boolean, };